- интеграция скрипта генерации грамматик в linux
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@12451 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/packages/scripts/trunk/python/run-bison.py
@@ -1,11 +1,8 @@
 import subprocess, os, argparse, re, sys
 
-def run_bison(yxPath, yPath, cppPath, name):
-    if sys.platform == 'win32':
-        procname = "set CYGWIN=nodosfilewarning& set BISON_PKGDATADIR=" + os.environ["BISON_PKGDATADIR"] + "&" + os.environ["BISON_PATH"] + " -p" + name + " \"" + yPath + "\" -o\"" + cppPath + "\""
-    else:
-        procname = "set BISON_PKGDATADIR=" + os.environ["BISON_PKGDATADIR"] + ";" + os.environ["BISON_PATH"] + " -p" + name + "\"" + yPath + "\" -o\"" + cppPath + "\""
-
+def run_bison(yxPath, yPath, cppPath, name, bison):
+    procname = bison + " -p" + name + " \"" + yPath + "\" -o\"" + cppPath + "\""
+    
     output, errr = subprocess.Popen(procname,stdout = subprocess.PIPE, stderr = subprocess.PIPE, shell = True).communicate()
 
     errr = errr.decode(codepage)
@@ -13,17 +10,18 @@ def run_bison(yxPath, yPath, cppPath, name):
     errr = errr.replace(yPath,yxPath)
     
     if sys.platform == 'win32':
-        yPath = yPath.replace("\\","\\\\")   # ну это какой-то грязный хак, конечно
+        yPath = yPath.replace("\\","\\\\")
         yxPath = yxPath.replace("\\","\\\\")
         errr = errr.replace(yPath,yxPath)
 
     if not errr.count("error"):
-        strGRAM = open(cppPath, encoding = codepage).read()
-        fin = open(cppPath, 'w', encoding = codepage)
-        fin.write( strGRAM.replace(yPath,yxPath) )
-        fin.close()
+        if os.path.exists(cppPath):
+            strGRAM = open(cppPath, encoding = codepage).read()
+            fin = open(cppPath, 'w', encoding = codepage)
+            fin.write( strGRAM.replace(yPath,yxPath) )
+            fin.close()
 
-    return re.sub(r"^([^:]*:)((\\\\?[^:\\]*)*):(([0-9]*).([0-9]*)-[0-9.]*)(.*)$", r"\1\2(\5)\7", errr, flags=re.MULTILINE)
+    return errr
 
 def main():
     parser = argparse.ArgumentParser(usage = argparse.SUPPRESS, description = "run bison twice for multipass compiler grammar files")
@@ -36,6 +34,7 @@ def main():
     parser.add_argument('-n2', type = str, default = '', help = "2nd output y name",   required = True)
     parser.add_argument('-o1', type = str, default = '', help = "1st output cpp file", required = True)
     parser.add_argument('-o2', type = str, default = '', help = "2nd output cpp file", required = True)
+    parser.add_argument('-bison', type = str, default = '', help = "bison path", required = True)
 
     args = parser.parse_args()
 
@@ -47,7 +46,7 @@ def main():
     yxPath  = os.path.abspath(args.inputFile)
     print(toolname + ": " + "parsing " + yPath)
     
-    out1 = run_bison(yxPath, yPath, cppPath, args.n1)
+    out1 = run_bison(yxPath, yPath, cppPath, args.n1, args.bison)
 
     cppPath = os.path.abspath(args.o2)
     if os.path.exists(cppPath):
@@ -56,7 +55,7 @@ def main():
     yxPath  = os.path.abspath(args.inputFile)
     print(toolname + ": " + "parsing " + yPath)
 
-    out2 = run_bison(yxPath, yPath, cppPath, args.n2)
+    out2 = run_bison(yxPath, yPath, cppPath, args.n2, args.bison)
 
     if out1 == out2:
         print(toolname + ": " + "bison output:")
@@ -68,15 +67,19 @@ def main():
             conflicts = file_conflict.read()
             conf_pos = conflicts[:conflicts.find("State ")].count("\n") + 1
             
-            out1 = out1.replace(": conflicts:", "(" + str(conf_pos) + ") : warning C0000:")
+            if sys.platform == 'win32':
+                out1 = out1.replace(": conflicts:", "(" + str(conf_pos) + ") : warning C0000:")
+            else:
+                out1 = out1.replace(": conflicts:", ":" + str(conf_pos) + ": warning:")
             conend = out1[:con].rfind("\n")
             if conend == -1:
                 conend = 0
             out1 = out1[:conend] + out1[conend:con].replace(yxPath, cppPath.replace(".cpp", ".output"), 1) + out1[con:]
 
-        out1 = out1.replace(": error:"," : error C0000:")
-        out1 = out1.replace("\\\\","\\")
-        print(out1)
+        if sys.platform == 'win32':
+            out1 = out1.replace(": error:"," : error C0000:")
+            out1 = out1.replace("\\\\","\\")
+        print(out1, file=sys.stderr)
         sys.exit(0)
     else:
         sys.exit(yxPath + ": error : bison console output doesn't match for grammar files")
